+++ src/tests/net-core/Xpath/AbstractXPathEngineTest.cs
@@ -22,16 +22,19 @@ namespace Org.XmlUnit.Xpath {
     public abstract class AbstractXPathEngineTest {
 
         private ISource source;
+        private XmlElement sourceRootElement;
 
         protected abstract IXPathEngine Engine { get; }
 
         [SetUp] public void ReadSource() {
             source = InputBuilder.FromFile(TestResources.BLAME_FILE).Build();
+            XmlDocument doc = new XmlDocument();
+            doc.Load(InputBuilder.FromFile(TestResources.BLAME_FILE).Build().Reader);
+            sourceRootElement = doc.DocumentElement;
         }
 
         [Test] public void SelectNodesWithNoMatches() {
-            IEnumerable<XmlNode> i = Engine.SelectNodes("foo",
-                                                                   source);
+            IEnumerable<XmlNode> i = Engine.SelectNodes("foo", source);
             Assert.IsNotNull(i);
             Assert.IsFalse(i.GetEnumerator().MoveNext());
         }
@@ -114,6 +117,96 @@ namespace Org.XmlUnit.Xpath {
             m["x"] = "urn:test:1";
             e.NamespaceContext = m;
             IEnumerable<XmlNode> it = e.SelectNodes("/x:d/x:e", source);
+            Assert.IsTrue(it.GetEnumerator().MoveNext());
+        }
+
+        [Test] public void SelectNodesWithNoMatches_NodeVersion() {
+            IEnumerable<XmlNode> i = Engine.SelectNodes("foo", sourceRootElement);
+            Assert.IsNotNull(i);
+            Assert.IsFalse(i.GetEnumerator().MoveNext());
+        }
+
+        [Test] public void SelectNodesWithSingleMatch_NodeVersion() {
+            IEnumerable<XmlNode> i = Engine.SelectNodes("//ul", sourceRootElement);
+            Assert.IsNotNull(i);
+            IEnumerator<XmlNode> it = i.GetEnumerator();
+            Assert.IsTrue(it.MoveNext());
+            Assert.AreEqual("ul", it.Current.Name);
+            Assert.IsFalse(it.MoveNext());
+        }
+
+        [Test] public void SelectNodesWithMultipleMatchs_NodeVersion() {
+            IEnumerable<XmlNode> i = Engine.SelectNodes("//li", sourceRootElement);
+            Assert.IsNotNull(i);
+            int count = 0;
+            foreach (XmlNode n in i) {
+                count++;
+                Assert.AreEqual("li", n.Name);
+            }
+            Assert.AreEqual(4, count);
+        }
+
+        [Test]
+        public void SelectNodesWithInvalidXPath_NodeVersion() {
+            try {
+                Engine.SelectNodes("//li[", sourceRootElement);
+                Assert.Fail("expected an exception");
+            } catch (XMLUnitException) {
+                // expected
+            }
+        }
+
+        [Test] public void EvaluateWithNoMatches_NodeVersion() {
+            Assert.AreEqual(string.Empty, Engine.Evaluate("foo", sourceRootElement));
+        }
+
+        [Test] public void EvaluateWithSingleMatch_NodeVersion() {
+            Assert.AreEqual("Don't blame it on the...",
+                            Engine.Evaluate("//title", sourceRootElement));
+        }
+
+        [Test] public void EvaluateWithSingleMatchTextSelector_NodeVersion() {
+            Assert.AreEqual("Don't blame it on the...",
+                            Engine.Evaluate("//title/text()", sourceRootElement));
+        }
+
+        [Test] public void EvaluateWithMultipleMatches_NodeVersion() {
+            Assert.AreEqual("sunshine",
+                            Engine.Evaluate("//li", sourceRootElement));
+        }
+
+        [Test]
+        public void EvaluateWithInvalidXPath_NodeVersion() {
+            try {
+                Engine.Evaluate("//li[", sourceRootElement);
+                Assert.Fail("expected an exception");
+            } catch (XMLUnitException) {
+                // expected
+            }
+        }
+
+        [Test] public void SelectNodesWithNS_NodeVersion() {
+            IXPathEngine e = Engine;
+            XmlDocument doc = new XmlDocument();
+            doc.Load(InputBuilder.FromString("<n:d xmlns:n='urn:test:1'><n:e/></n:d>").Build().Reader);
+            sourceRootElement = doc.DocumentElement;
+            Dictionary<string, string> m = new Dictionary<string, string>();
+            m["x"] = "urn:test:1";
+            e.NamespaceContext = m;
+            IEnumerable<XmlNode> it = e.SelectNodes("/x:d/x:e", sourceRootElement);
+            Assert.IsTrue(it.GetEnumerator().MoveNext());
+        }
+
+        [Test] public void SelectNodesWithDefaultNS_NodeVersion() {
+            IXPathEngine e = Engine;
+            XmlDocument doc = new XmlDocument();
+            doc.Load(InputBuilder.FromString("<d xmlns='urn:test:1'><e/></d>")
+                     .Build().Reader);
+            sourceRootElement = doc.DocumentElement;
+            Dictionary<string, string> m = new Dictionary<string, string>();
+            m["x"] = "urn:test:1";
+            e.NamespaceContext = m;
+            IEnumerable<XmlNode> it = e.SelectNodes("/x:d/x:e", sourceRootElement);
             Assert.IsTrue(it.GetEnumerator().MoveNext());
         }