+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
   XMLUnit.Constraints
 * The XMLUnit.Constraints nuget package now depends on NUNit 2.6.4 -
   which it has been compiled against - rather than 2.5.10.
+* added new overloads to `IXPathEngine`
 
 ## XMLUnit.NET 2.0.0-alpha-02 - /Released 2015-11-21/